+++ configs/shapenet_cars.txt
@@ -4,3 +4,10 @@ split_file = datasets/shapenet/data/split_cars.npz
 input_data_glob = /*/model.obj
 sample_std_dev = [0.08, 0.02, 0.003]
 sample_ratio = [0.01, 0.49, 0.5]
+
+input_res = 256
+pc_samples = 3000
+index = 1
+num_points=1000
+cam_position=[0, 1, 0]
+cam_orientation=[90.0, 0.0, 180.0]
